+++ libraries/Bluefruit52Lib/src/bluefruit.h
@@ -74,6 +74,7 @@
 #include "services/BLEHidGeneric.h"
 #include "services/BLEHidAdafruit.h"
 #include "services/BLEMidi.h"
+#include "services/EddyStone.h"
 
 #include "clients/BLEAncs.h"
 #include "clients/BLEClientUart.h"