+++ changelog.md
@@ -37,6 +37,7 @@
 - Add the option to advertise for a specific time. There are multiple timeouts: initial fast advertising mode, slow advertising mode, and an optional delay to stop advertising entirely. Values can be set in multiple of timeout ticks (0.625ms per unit) or in ms (approximate since it gets converted to 0.625ms units). The optional timeout to stop advertising entire is set via the `.start(timeout)` parameter.
 - Blue LED will blink 2x in fast mode than slow mode.
 - Default timeout to slow mode = 30 seconds. Default adv intervals are fast mode = 20 ms, slow = 152.5 ms
+- Expand addUuid(), addService() to take a list of UUID
 - Add new example `advance_adv`
 
 ### BLEGAP (To support multiple connections, etc.)